Refactor helper functions file
