+++ e2e/helpers/generated_order_helper.js
@@ -41,5 +41,4 @@ module.exports = {
         break;
     }
   },
-
 };
